Split declaration precollection task tests
