+++ src/typechecker/tests/declaration_validation.rs
@@ -1,5 +1,6 @@
 use super::*;
 
+mod precollection_tasks;
 mod resolver_replay;
 mod tasks;
