+++ src/typechecker/tests/declaration_validation/precollection_tasks.rs
@@ -1,0 +1,162 @@
+use super::*;
+
+#[test]
+fn self_type_context_validation_tasks_collect_declarations() {
+    let program = parse_program(
+        r#"
+Point: { x: i32 }
+Option<T>: Some(T), None
+
+Json: behavior {
+    encode: (Self) StaticString
+}
+
+Pretty.extends(Json)
+
+make = () Point { Point { x: 1 } }
+
+Point.get = (self: Point) i32 { self.x }
+
+Point.impl = {
+    x_value = (self: Point) i32 { self.x }
+}
+
+Point.requires(Json)
+
+result := 1
+"#,
+    );
+
+    let tasks = TypeChecker::collect_self_type_context_validation_tasks(&program.declarations);
+
+    assert_eq!(tasks.len(), 9);
+    assert!(matches!(
+        tasks[0],
+        SelfTypeContextValidationTask::Struct { .. }
+    ));
+    assert!(matches!(
+        tasks[1],
+        SelfTypeContextValidationTask::Enum { .. }
+    ));
+    assert!(matches!(
+        tasks[2],
+        SelfTypeContextValidationTask::Behavior { .. }
+    ));
+    assert!(matches!(
+        tasks[3],
+        SelfTypeContextValidationTask::BehaviorExtends { .. }
+    ));
+    assert!(matches!(
+        tasks[4],
+        SelfTypeContextValidationTask::Function { .. }
+    ));
+    assert!(matches!(
+        tasks[5],
+        SelfTypeContextValidationTask::Method { .. }
+    ));
+    assert!(matches!(
+        tasks[6],
+        SelfTypeContextValidationTask::ImplBlock { .. }
+    ));
+    assert!(matches!(
+        tasks[7],
+        SelfTypeContextValidationTask::Requires { .. }
+    ));
+    assert!(matches!(
+        tasks[8],
+        SelfTypeContextValidationTask::TopLevelExpr { .. }
+    ));
+}
+
+#[test]
+fn ast_precollection_validation_tasks_collect_self_and_extends_work() {
+    let program = parse_program(
+        r#"
+Point: { x: i32 }
+
+Json: behavior {
+    encode: (Self) StaticString
+}
+
+Pretty.extends(Json)
+
+Point.impl = {
+    x_value = (self: Point) i32 { self.x }
+}
+
+result := 1
+"#,
+    );
+
+    let tasks = TypeChecker::collect_ast_precollection_validation_tasks(&program.declarations);
+
+    assert_eq!(tasks.self_type_contexts.len(), 5);
+    assert_eq!(tasks.behavior_associations.extends.len(), 1);
+    assert_eq!(tasks.behavior_associations.extends[0].behavior, "Pretty");
+    assert_eq!(tasks.behavior_associations.extends[0].parent, "Json");
+    assert!(tasks.behavior_associations.impls.is_empty());
+    assert!(tasks.behavior_associations.requires.is_empty());
+}
+
+#[test]
+fn ast_declaration_collection_tasks_include_precollection_validation_work() {
+    let program = parse_program(
+        r#"
+Point: { x: i32 }
+
+Json: behavior {
+    encode: (Self) StaticString
+}
+
+Pretty.extends(Json)
+
+make = () Point { Point { x: 1 } }
+"#,
+    );
+
+    let tasks = TypeChecker::collect_ast_declaration_collection_tasks(&program.declarations);
+
+    assert_eq!(tasks.types.len(), 1);
+    assert_eq!(tasks.behaviors.len(), 1);
+    assert_eq!(tasks.callable.len(), 1);
+    assert_eq!(tasks.precollection_validations.self_type_contexts.len(), 4);
+    assert_eq!(
+        tasks
+            .precollection_validations
+            .behavior_associations
+            .extends
+            .len(),
+        1
+    );
+}
+
+#[test]
+fn ast_declaration_collection_bundle_replays_collection_passes() {
+    let program = parse_program(
+        r#"
+Point: { x: i32 }
+
+Json: behavior {
+    encode: (Self) StaticString
+}
+
+make = () Point { Point { x: 1 } }
+
+Point.get = (self: Point) i32 { self.x }
+
+Point.impl = {
+    x_value = (self: Point) i32 { self.x }
+}
+"#,
+    );
+    let tasks = TypeChecker::collect_ast_declaration_collection_tasks(&program.declarations);
+    let mut checker = TypeChecker::new();
+
+    checker.collect_ast_declarations_from_tasks(&tasks);
+
+    assert!(checker.structs.contains_key("Point"));
+    assert!(checker.behaviors.contains_key("Json"));
+    assert!(checker.functions.contains_key("make"));
+    assert!(checker.methods.contains_key("Point.get"));
+    assert!(checker.methods.contains_key("Point.x_value"));
+}
